Added custom hooks in the NavBar, Sidebar, and SearchBar
