--- src/shared/ui/SearchBar/hooks/useSearchQuery.tsx
@@ -1,1 +1,0 @@
-const useSearchQuery = () => {};
