+++ src/shared/ui/SearchBar/types/types.tsx
@@ -1,0 +1,6 @@
+export type Flashcard = {
+  id: string;
+  title: string;
+  description: string;
+  lastUpdated: string;
+};
